An atom based identity.
A major question lingering over the atomic microstates refactor is how
it would work with identity (the mechanism for maintaing a stable
graph of references).

This change contains an implementation of identity using the atomic
microstates architecture.

The following new constructs were required:

** Polymorphic `atomOf()` **

With an atomic microstate architecture, microstates no longer hold a
reference directly to their values. Instead, they hold a reference
directly to their paths, and in order to retrieve their values, they
look up that their path inside of an atom.

This poses a problem for stable identities because the atom needs to
change, but the JavaScript reference must remain the same. To solve
this, we make `atomOf()` a polymorhpic function of the `AtomOf`
typeclass.

The default implementation of `atomOf` returns the atom to which the
current microstate holds a reference. The `Id<Type>` implementation
however, returns the top level value of the identity which changes
over time with each transition.

** Extended MicrostateType() **

There is a lot that is the same between the base microstate type
instantiated with `create`, and the the Identity microstate
type. Things like maintaining the meta information, the `set`
transition, caching instance properties, etc...

However, there are some important places where we want to diverge,
specifically in how transitions are handled, and properties are
resolved.

This introduces new parameters to the `transitionFn` and `propertyFn`,
specifically, the `Type`, and `path` parameters where the transition
is happening or the property is being resolved are both now passed.
is called, or a property is resolved. This allows us to have a single.

** Tree **

In order to reason about the entire tree as a whole, we introduce a
lightweight `Tree` class not dissimilar to the `Profuctor` class in
the current implementation.

It is expected that this will only be needed for the identity map, and
not for implementing the rest of the architecture.

** Higher Order Store **

Identity was previously interwined with the concept of `Store`. This
meant that integrating with other stores like Redux and Obvservable,
required a little bit of massaging or integrating with the observation
mechanism.

The Identity() function implemented here takes a microstate and a
transition event callback, returns two things: a function `get` which returns the
current reference to the root id microstate, and a `transition`
function which actually causes a transition to happen.

So, when a caller invokes a transition from the microstate graph, it
invokes the callback with the `Type`, `path`, `name`, `args` of where
the transition happened. This is the same signature as the transition
callback, and so the idea is that at some point, you invoke it with
those arguments.

So why would you want to split a transition into two phases? The
answer is that by makeing it

So for example, a fully synchronous Store would immediately delegate
from the callback to the transition function:

```js
let { get, transition } = Identity(create(Number, 0), (...args) => transition(...args));
id.get().increment();
id.get() //> Id<Number> { 1 }
```

Redux, which is also a synchronous store can be implementing in a
straightforward fashion

```js
let { get, transition } = Identity(Create(Number, 0), (Type, path, name, args) => {
  store.dispatch({
    isMicrostateTransition: true
    type: `${Type.name}.#{name}${path.join('/')}`,
    Type,
    path,
    name,
    args
  })
  return get()
})

// and in reducer:
function microstateRducer(state, action) {
  if (action.isMicrostateTransition) {
    let { Type, path, name, args } = action;
    transition(Type, path, name, args);
    return id.get();
  }
}

```

Obvioeusly, the current store is implemented easily too:

```js
export default function Store(Type, value, observe = x => x) {
  let { get, transition } = Identity(Type, value, (...args) => observe(transition(...args)));
  return get();
}
```

Follow on work:

- Make sure that microstate types are stable: same constructor for
  given `Type`
- make sure all function parameters make sense and are aligned
- a uniform linking function (link / link2 is a no-go)