valueOf() should work with any object.
A pillar of microstates is that you can ask the `valueOf` anything at
any time and it should return to you a simple POJO which can be used
to reconstruct the value.

For microstates this is the underlying value, for everything other
thing, it's the thing.

This was blowing up before because non-microstates don't have `Meta`
information.