Add an example of a foreign key relationship.
This adds a `DB` example that contains a number of tables each of
which have rows that link to each other. This uses the ability to link
anywhere in the store to create a factory method that creates entities
elswhere in the store.

In this example, the `Blog` record has a reference to a `Person` which
is the author. The factory for author attribute of a blog, creates
that person in the database, and then returns that person.