Ignore examples and query in coverage report.
We definitely want to have a solid coverage of the core library being
developed under lab, however, the examples directory is not included
in that core library. Also, the query library was copied over
wholesale from microstates as a convenience.

This removes examples and query.js from the coverage report.